feat: Add a script that sets up your web servers for the deployment of web_static

+++ 0-setup_web_static.sh
@@ -2,7 +2,7 @@
 # Install Nginx if it not already installed
 
 sudo apt update -y
-sudo apt upgrade -y
+# sudo apt upgrade -y
 sudo apt install nginx -y
 
 # Creates the folder /data/web_static/shared/ if it doesn’t already exist
@@ -16,14 +16,18 @@ content="<html><head></head><body>Holberton School</body></html>"
 echo "$content" > /data/web_static/releases/test/index.html
 
 # Creates a symbolic link /data/web_static/current linked to the /data/web_static/releases/test/ folder.
-sudo rm /data/web_static/current
+if [ -L "/data/web_static/current" ];then
+    sudo rm /data/web_static/current
+fi
 sudo ln -sf /data/web_static/releases/test/ /data/web_static/current
 
 # Give ownership of the /data/ folder to the ubuntu user AND group .
 sudo chown -hR ubuntu:ubuntu /data/
 
 # Update the Nginx configuration to serve the content of /data/web_static/current/ to hbnb_static
-content=$"\n\tlocation /hbnb_static {\n\t\talias /data/web_static/current/;\n\t}"
-sudo sed -i "/server_name _;/a\\$content" /etc/nginx/sites-available/default
+if ! grep -q "location /hbnb_static {" /etc/nginx/sites-available/default; then
+    content=$"\n\tlocation /hbnb_static {\n\t\talias /data/web_static/current/;\n\t}"
+    sudo sed -i "/server_name _;/a\\$content" /etc/nginx/sites-available/default
+fi
 # reload nginx
 sudo service nginx restart
